// src/types.ts
export type Address =`0x${string}`;
export type Hex = `0x${string}`;
export type Hash = `0x${string}`;
export type EmptyHex = `0x`;
export type PromiseOrValue<T> = T | Promise<T>;
export type BigNumberish = string | bigint | number;
export type BytesLike = Uint8Array | string;


export interface SendUserOpResponse {
	userOpHash: Hash,
	transactionHash?: Hash 
}

export interface ClientOptions {
	chainId: number;
	privateKey?: string;
	rpcUrl?: string;
	isSponsoredTrx?: boolean
	noAccountCreateSponsorship?: boolean
	paymasterEndpoint?: string;
	paymaster?: "PIMLICO" | "STACKUP" | "MAGMAR";
	paymasterId?: string; //for magmar paymaster
	erc20token?: string;
	bundlerEndpoint?:string;
}

export interface TelegramClientOptions {
	chainId: number;
	telegramUserId: string;
	walletAddress: Address;
	rpcUrl?: string;
	isSponsoredTrx?: boolean
	paymasterEndpoint?: string;
	paymaster?: "PIMLICO" | "STACKUP",
	bundlerEndpoint?:string;
}

export interface UserOperationCallData {
	target: Address;
	data: Hex;
	value?: bigint;
}

export type BatchUserOperationCallData = UserOperationCallData[];
export interface UserOperationRequest {
	/* the origin of the request */
	sender: Address;
	/* nonce (as hex) of the transaction, returned from the entrypoint for this Address */
	nonce: Hex;
	/* the initCode for creating the sender if it does not exist yet, otherwise "0x" */
	initCode: Hex | EmptyHex;
	/* the callData passed to the target */
	callData: Hex;
	/* Gas value (as hex) used by inner account execution */
	callGasLimit: Hex;
	/* Actual gas (as hex) used by the validation of this UserOperation */
	verificationGasLimit: Hex;
	/* Gas overhead (as hex) of this UserOperation */
	preVerificationGas: Hex;
	/* Maximum fee per gas (similar to EIP-1559 max_fee_per_gas) (as hex)*/
	maxFeePerGas: Hex;
	/* Maximum priority fee per gas (similar to EIP-1559 max_priority_fee_per_gas) (as hex)*/
	maxPriorityFeePerGas: Hex;
	/* Address of paymaster sponsoring the transaction, followed by extra data to send to the paymaster ("0x" for self-sponsored transaction) */
	paymasterAndData: Hex | EmptyHex;
	/* Data passed into the account along with the nonce during the verification step */
	signature: Hex;
  }